+++ index.js
@@ -29,7 +29,7 @@ client.on('messageCreate', msg =>{
   if(msg.author.id == client.user.id) return;
   if(!msg.content.startsWith(prefix)) return;
 
-  msg.reply({content: msg.content, embeds: [embeds.example(msg.content)]})
+  msg.reply({embeds: [embeds.playerGuess("1", "⬛⬛🟩⬛🟨\n🟨🟨⬛🟩⬛\n⬛🟨🟩⬛🟩\n🟩⬛🟨🟨🟨\n🟩🟩🟩🟨🟨")]})
   console.log(msg.author.tag + " - " + msg.content);
   msg.react('✅');
 })