+++ src/modules/embeds.js
@@ -1,10 +1,14 @@
 const { MessageEmbed } = require("discord.js");
 
-const example = (msgContent) => {
+const playerGuess = (id, guess) => {
   return new MessageEmbed()
-    .setTitle("Test")
-    .setDescription(msgContent);
+  .setTitle("Your wordle #"+id)
+	.setURL('https://www.powerlanguage.co.uk/wordle/')
+  .setAuthor({ name: "Discord Wordle", iconURL: "", url: "https://github.com/nevah5/wordlebot"})
+  .addField("Your guesses:", guess, true)
+  .setFooter({text: "1 guess left, guess with /guess <YOUR_GUESS>"})
+  .setColor("#6AAA64");
 }
 module.exports = {
-  example
+  playerGuess
 };